+++ KriptonAndroidLibrary/src/test/java/sqlite/feature/contentprovider/kripton35/nolog/Person2DAOImpl.java
@@ -53,6 +53,14 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
 
   private SQLiteStatement insertNamePreparedStatement1;
 
+  private SQLiteStatement deleteRawPreparedStatement2;
+
+  private SQLiteStatement deleteBeanPreparedStatement3;
+
+  private SQLiteStatement updateRawPreparedStatement4;
+
+  private SQLiteStatement updateBeanPreparedStatement5;
+
   public Person2DAOImpl(BindPerson2DataSource dataSet) {
     super(dataSet);
   }
@@ -79,7 +87,7 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
    */
   @Override
   public void insertBean(Person bean) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     _contentValues.put("city", bean.city);
     if (bean.birthCity!=null) {
       _contentValues.put("birth_city", bean.birthCity);
@@ -172,7 +180,7 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
    */
   @Override
   public void insertName(String tempName) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
 
     if (tempName!=null) {
       _contentValues.put("name", tempName);
@@ -256,24 +264,27 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
    */
   @Override
   public int deleteRaw(long id) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     _contentValues.addWhereArgs(String.valueOf(id));
 
-    StringBuilder _sqlBuilder=getSQLStringBuilder();
     // generation CODE_001 -- BEGIN
     // generation CODE_001 -- END
+    if (deleteRawPreparedStatement2==null) {
+      StringBuilder _sqlBuilder=getSQLStringBuilder();
 
-    // manage WHERE arguments -- BEGIN
+      // manage WHERE arguments -- BEGIN
 
-    // manage WHERE statement
-    String _sqlWhereStatement=" id = ?";
-    _sqlBuilder.append(_sqlWhereStatement);
+      // manage WHERE statement
+      String _sqlWhereStatement=" id = ?";
+      _sqlBuilder.append(_sqlWhereStatement);
 
-    // manage WHERE arguments -- END
+      // manage WHERE arguments -- END
 
-    // generate sql
-    String _sql="DELETE FROM person WHERE id = ?";
-    int result = KriptonDatabaseWrapper.updateDelete(dataSource, _sql, _contentValues);
+      // generate sql
+      String _sql="DELETE FROM person WHERE id = ?";
+      deleteRawPreparedStatement2 = KriptonDatabaseWrapper.compile(dataSource, _sql);
+    }
+    int result = KriptonDatabaseWrapper.updateDelete(dataSource, deleteRawPreparedStatement2, _contentValues);
     return result;
   }
 
@@ -351,16 +362,16 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
    */
   @Override
   public boolean deleteRaw(long id, String where, String[] args) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     _contentValues.addWhereArgs(String.valueOf(id));
 
-    StringBuilder _sqlBuilder=getSQLStringBuilder();
     // generation CODE_001 -- BEGIN
     // initialize dynamic where
     String _sqlDynamicWhere=where;
     // initialize dynamic where args
     String[] _sqlDynamicWhereArgs=args;
     // generation CODE_001 -- END
+    StringBuilder _sqlBuilder=getSQLStringBuilder();
 
     // manage WHERE arguments -- BEGIN
 
@@ -454,24 +465,27 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
    */
   @Override
   public long deleteBean(Person bean) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     _contentValues.addWhereArgs(String.valueOf(bean.id));
 
-    StringBuilder _sqlBuilder=getSQLStringBuilder();
     // generation CODE_001 -- BEGIN
     // generation CODE_001 -- END
+    if (deleteBeanPreparedStatement3==null) {
+      StringBuilder _sqlBuilder=getSQLStringBuilder();
 
-    // manage WHERE arguments -- BEGIN
+      // manage WHERE arguments -- BEGIN
 
-    // manage WHERE statement
-    String _sqlWhereStatement=" id = ?";
-    _sqlBuilder.append(_sqlWhereStatement);
+      // manage WHERE statement
+      String _sqlWhereStatement=" id = ?";
+      _sqlBuilder.append(_sqlWhereStatement);
 
-    // manage WHERE arguments -- END
+      // manage WHERE arguments -- END
 
-    // generate sql
-    String _sql="DELETE FROM person WHERE id = ?";
-    int result = KriptonDatabaseWrapper.updateDelete(dataSource, _sql, _contentValues);
+      // generate sql
+      String _sql="DELETE FROM person WHERE id = ?";
+      deleteBeanPreparedStatement3 = KriptonDatabaseWrapper.compile(dataSource, _sql);
+    }
+    int result = KriptonDatabaseWrapper.updateDelete(dataSource, deleteBeanPreparedStatement3, _contentValues);
     return result;
   }
 
@@ -543,7 +557,7 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
    */
   @Override
   public int updateRaw(String name, long id) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     if (name!=null) {
       _contentValues.put("name", name);
     } else {
@@ -552,33 +566,24 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
 
     _contentValues.addWhereArgs(String.valueOf(id));
 
-    StringBuilder _sqlBuilder=getSQLStringBuilder();
     // generation CODE_001 -- BEGIN
     // generation CODE_001 -- END
+    if (updateRawPreparedStatement4==null) {
+      StringBuilder _sqlBuilder=getSQLStringBuilder();
 
-    // manage WHERE arguments -- BEGIN
+      // manage WHERE arguments -- BEGIN
 
-    // manage WHERE statement
-    String _sqlWhereStatement=" id=?";
-    _sqlBuilder.append(_sqlWhereStatement);
+      // manage WHERE statement
+      String _sqlWhereStatement=" id=?";
+      _sqlBuilder.append(_sqlWhereStatement);
 
-    // manage WHERE arguments -- END
+      // manage WHERE arguments -- END
 
-    // generate sql
-    String _sql="UPDATE person SET name=? WHERE id=?";
-
-    // log for content values -- BEGIN
-    Triple<String, Object, KriptonContentValues.ParamType> _contentValue;
-    for (int i = 0; i < _contentValues.size(); i++) {
-      _contentValue = _contentValues.get(i);
-      if (_contentValue.value1==null) {
-        Logger.info("==> :%s = <null>", _contentValue.value0);
-      } else {
-        Logger.info("==> :%s = '%s' (%s)", _contentValue.value0, StringUtils.checkSize(_contentValue.value1), _contentValue.value1.getClass().getCanonicalName());
-      }
+      // generate sql
+      String _sql="UPDATE person SET name=? WHERE id=?";
+      updateRawPreparedStatement4 = KriptonDatabaseWrapper.compile(dataSource, _sql);
     }
-    // log for content values -- END
-    int result = KriptonDatabaseWrapper.updateDelete(dataSource, _sql, _contentValues);
+    int result = KriptonDatabaseWrapper.updateDelete(dataSource, updateRawPreparedStatement4, _contentValues);
     return result;
   }
 
@@ -678,7 +683,7 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
    */
   @Override
   public int updateRaw(String name, long id, String where) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     if (name!=null) {
       _contentValues.put("name", name);
     } else {
@@ -687,11 +692,11 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
 
     _contentValues.addWhereArgs(String.valueOf(id));
 
-    StringBuilder _sqlBuilder=getSQLStringBuilder();
     // generation CODE_001 -- BEGIN
     // initialize dynamic where
     String _sqlDynamicWhere=where;
     // generation CODE_001 -- END
+    StringBuilder _sqlBuilder=getSQLStringBuilder();
 
     // manage WHERE arguments -- BEGIN
 
@@ -703,18 +708,6 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
 
     // generate sql
     String _sql=String.format("UPDATE person SET name=? WHERE id=?%s", StringUtils.ifNotEmptyAppend(_sqlDynamicWhere," AND "));
-
-    // log for content values -- BEGIN
-    Triple<String, Object, KriptonContentValues.ParamType> _contentValue;
-    for (int i = 0; i < _contentValues.size(); i++) {
-      _contentValue = _contentValues.get(i);
-      if (_contentValue.value1==null) {
-        Logger.info("==> :%s = <null>", _contentValue.value0);
-      } else {
-        Logger.info("==> :%s = '%s' (%s)", _contentValue.value0, StringUtils.checkSize(_contentValue.value1), _contentValue.value1.getClass().getCanonicalName());
-      }
-    }
-    // log for content values -- END
     int result = KriptonDatabaseWrapper.updateDelete(dataSource, _sql, _contentValues);
     return result;
   }
@@ -817,7 +810,7 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
    */
   @Override
   public int updateRaw(String name, long id, String where, String[] args) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     if (name!=null) {
       _contentValues.put("name", name);
     } else {
@@ -826,13 +819,13 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
 
     _contentValues.addWhereArgs(String.valueOf(id));
 
-    StringBuilder _sqlBuilder=getSQLStringBuilder();
     // generation CODE_001 -- BEGIN
     // initialize dynamic where
     String _sqlDynamicWhere=where;
     // initialize dynamic where args
     String[] _sqlDynamicWhereArgs=args;
     // generation CODE_001 -- END
+    StringBuilder _sqlBuilder=getSQLStringBuilder();
 
     // manage WHERE arguments -- BEGIN
 
@@ -849,18 +842,6 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
 
     // generate sql
     String _sql=String.format("UPDATE person SET name=? WHERE id=?%s", StringUtils.ifNotEmptyAppend(_sqlDynamicWhere," AND "));
-
-    // log for content values -- BEGIN
-    Triple<String, Object, KriptonContentValues.ParamType> _contentValue;
-    for (int i = 0; i < _contentValues.size(); i++) {
-      _contentValue = _contentValues.get(i);
-      if (_contentValue.value1==null) {
-        Logger.info("==> :%s = <null>", _contentValue.value0);
-      } else {
-        Logger.info("==> :%s = '%s' (%s)", _contentValue.value0, StringUtils.checkSize(_contentValue.value1), _contentValue.value1.getClass().getCanonicalName());
-      }
-    }
-    // log for content values -- END
     int result = KriptonDatabaseWrapper.updateDelete(dataSource, _sql, _contentValues);
     return result;
   }
@@ -967,7 +948,7 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
    */
   @Override
   public int updateBean(Person person) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     _contentValues.put("alias_parent_id", person.parentId);
     _contentValues.put("city", person.city);
     if (person.birthCity!=null) {
@@ -994,33 +975,24 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
 
     _contentValues.addWhereArgs(String.valueOf(person.id));
 
-    StringBuilder _sqlBuilder=getSQLStringBuilder();
     // generation CODE_001 -- BEGIN
     // generation CODE_001 -- END
+    if (updateBeanPreparedStatement5==null) {
+      StringBuilder _sqlBuilder=getSQLStringBuilder();
 
-    // manage WHERE arguments -- BEGIN
+      // manage WHERE arguments -- BEGIN
 
-    // manage WHERE statement
-    String _sqlWhereStatement=" id=?";
-    _sqlBuilder.append(_sqlWhereStatement);
+      // manage WHERE statement
+      String _sqlWhereStatement=" id=?";
+      _sqlBuilder.append(_sqlWhereStatement);
 
-    // manage WHERE arguments -- END
+      // manage WHERE arguments -- END
 
-    // generate sql
-    String _sql="UPDATE person SET alias_parent_id=?, city=?, birth_city=?, birth_day=?, value=?, name=?, surname=? WHERE id=?";
-
-    // log for content values -- BEGIN
-    Triple<String, Object, KriptonContentValues.ParamType> _contentValue;
-    for (int i = 0; i < _contentValues.size(); i++) {
-      _contentValue = _contentValues.get(i);
-      if (_contentValue.value1==null) {
-        Logger.info("==> :%s = <null>", _contentValue.value0);
-      } else {
-        Logger.info("==> :%s = '%s' (%s)", _contentValue.value0, StringUtils.checkSize(_contentValue.value1), _contentValue.value1.getClass().getCanonicalName());
-      }
+      // generate sql
+      String _sql="UPDATE person SET alias_parent_id=?, city=?, birth_city=?, birth_day=?, value=?, name=?, surname=? WHERE id=?";
+      updateBeanPreparedStatement5 = KriptonDatabaseWrapper.compile(dataSource, _sql);
     }
-    // log for content values -- END
-    int result = KriptonDatabaseWrapper.updateDelete(dataSource, _sql, _contentValues);
+    int result = KriptonDatabaseWrapper.updateDelete(dataSource, updateBeanPreparedStatement5, _contentValues);
     return result;
   }
 
@@ -1807,6 +1779,22 @@ public class Person2DAOImpl extends AbstractDao implements Person2DAO {
     if (insertNamePreparedStatement1!=null) {
       insertNamePreparedStatement1.close();
       insertNamePreparedStatement1=null;
+    }
+    if (deleteRawPreparedStatement2!=null) {
+      deleteRawPreparedStatement2.close();
+      deleteRawPreparedStatement2=null;
+    }
+    if (deleteBeanPreparedStatement3!=null) {
+      deleteBeanPreparedStatement3.close();
+      deleteBeanPreparedStatement3=null;
+    }
+    if (updateRawPreparedStatement4!=null) {
+      updateRawPreparedStatement4.close();
+      updateRawPreparedStatement4=null;
+    }
+    if (updateBeanPreparedStatement5!=null) {
+      updateBeanPreparedStatement5.close();
+      updateBeanPreparedStatement5=null;
     }
   }
 }
