+++ KriptonAndroidLibrary/src/test/java/sqlite/feature/foreignkeyaction/AlbumDaoImpl.java
@@ -21,7 +21,11 @@ import java.util.List;
  *  @see AlbumTable
  */
 public class AlbumDaoImpl extends AbstractDao implements AlbumDao {
-  private SQLiteStatement insertPreparedStatement0;
+  private SQLiteStatement updatePreparedStatement0;
+
+  private SQLiteStatement insertPreparedStatement1;
+
+  private SQLiteStatement deleteByIdPreparedStatement2;
 
   public AlbumDaoImpl(BindArtistDataSource dataSet) {
     super(dataSet);
@@ -184,7 +188,7 @@ public class AlbumDaoImpl extends AbstractDao implements AlbumDao {
    */
   @Override
   public long update(Album bean) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     _contentValues.put("artist_id", bean.artistId);
     if (bean.name!=null) {
       _contentValues.put("name", bean.name);
@@ -194,20 +198,23 @@ public class AlbumDaoImpl extends AbstractDao implements AlbumDao {
 
     _contentValues.addWhereArgs(String.valueOf(bean.id));
 
-    StringBuilder _sqlBuilder=getSQLStringBuilder();
     // generation CODE_001 -- BEGIN
     // generation CODE_001 -- END
+    if (updatePreparedStatement0==null) {
+      StringBuilder _sqlBuilder=getSQLStringBuilder();
 
-    // manage WHERE arguments -- BEGIN
+      // manage WHERE arguments -- BEGIN
 
-    // manage WHERE statement
-    String _sqlWhereStatement=" id=?";
-    _sqlBuilder.append(_sqlWhereStatement);
+      // manage WHERE statement
+      String _sqlWhereStatement=" id=?";
+      _sqlBuilder.append(_sqlWhereStatement);
 
-    // manage WHERE arguments -- END
+      // manage WHERE arguments -- END
 
-    // generate sql
-    String _sql="UPDATE album SET artist_id=?, name=? WHERE id=?";
+      // generate sql
+      String _sql="UPDATE album SET artist_id=?, name=? WHERE id=?";
+      updatePreparedStatement0 = KriptonDatabaseWrapper.compile(dataSource, _sql);
+    }
 
     // display log
     Logger.info("UPDATE album SET artist_id=:artistId, name=:name WHERE id=?");
@@ -230,7 +237,7 @@ public class AlbumDaoImpl extends AbstractDao implements AlbumDao {
       Logger.info("==> param%s: '%s'",(_whereParamCounter++), StringUtils.checkSize(_whereParamItem));
     }
     // log for where parameters -- END
-    int result = KriptonDatabaseWrapper.updateDelete(dataSource, _sql, _contentValues);
+    int result = KriptonDatabaseWrapper.updateDelete(dataSource, updatePreparedStatement0, _contentValues);
     return result;
   }
 
@@ -253,7 +260,7 @@ public class AlbumDaoImpl extends AbstractDao implements AlbumDao {
    */
   @Override
   public long insert(Album bean) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     _contentValues.put("artist_id", bean.artistId);
     if (bean.name!=null) {
       _contentValues.put("name", bean.name);
@@ -286,12 +293,12 @@ public class AlbumDaoImpl extends AbstractDao implements AlbumDao {
     // log for insert -- END 
 
     // insert operation
-    if (insertPreparedStatement0==null) {
+    if (insertPreparedStatement1==null) {
       // generate SQL for insert
       String _sql=String.format("INSERT INTO album (%s) VALUES (%s)", _contentValues.keyList(), _contentValues.keyValueList());
-      insertPreparedStatement0 = KriptonDatabaseWrapper.compile(dataSource, _sql);
+      insertPreparedStatement1 = KriptonDatabaseWrapper.compile(dataSource, _sql);
     }
-    long result = KriptonDatabaseWrapper.insert(dataSource, insertPreparedStatement0, _contentValues);
+    long result = KriptonDatabaseWrapper.insert(dataSource, insertPreparedStatement1, _contentValues);
     bean.id=result;
 
     return result;
@@ -314,23 +321,26 @@ public class AlbumDaoImpl extends AbstractDao implements AlbumDao {
    */
   @Override
   public long deleteById(long id) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     _contentValues.addWhereArgs(String.valueOf(id));
 
-    StringBuilder _sqlBuilder=getSQLStringBuilder();
     // generation CODE_001 -- BEGIN
     // generation CODE_001 -- END
+    if (deleteByIdPreparedStatement2==null) {
+      StringBuilder _sqlBuilder=getSQLStringBuilder();
 
-    // manage WHERE arguments -- BEGIN
+      // manage WHERE arguments -- BEGIN
 
-    // manage WHERE statement
-    String _sqlWhereStatement=" id=?";
-    _sqlBuilder.append(_sqlWhereStatement);
+      // manage WHERE statement
+      String _sqlWhereStatement=" id=?";
+      _sqlBuilder.append(_sqlWhereStatement);
 
-    // manage WHERE arguments -- END
+      // manage WHERE arguments -- END
 
-    // generate sql
-    String _sql="DELETE FROM album WHERE id=?";
+      // generate sql
+      String _sql="DELETE FROM album WHERE id=?";
+      deleteByIdPreparedStatement2 = KriptonDatabaseWrapper.compile(dataSource, _sql);
+    }
 
     // display log
     Logger.info("DELETE FROM album WHERE id=?");
@@ -341,14 +351,22 @@ public class AlbumDaoImpl extends AbstractDao implements AlbumDao {
       Logger.info("==> param%s: '%s'",(_whereParamCounter++), StringUtils.checkSize(_whereParamItem));
     }
     // log for where parameters -- END
-    int result = KriptonDatabaseWrapper.updateDelete(dataSource, _sql, _contentValues);
+    int result = KriptonDatabaseWrapper.updateDelete(dataSource, deleteByIdPreparedStatement2, _contentValues);
     return result;
   }
 
   public void clearCompiledStatements() {
-    if (insertPreparedStatement0!=null) {
-      insertPreparedStatement0.close();
-      insertPreparedStatement0=null;
+    if (updatePreparedStatement0!=null) {
+      updatePreparedStatement0.close();
+      updatePreparedStatement0=null;
+    }
+    if (insertPreparedStatement1!=null) {
+      insertPreparedStatement1.close();
+      insertPreparedStatement1=null;
+    }
+    if (deleteByIdPreparedStatement2!=null) {
+      deleteByIdPreparedStatement2.close();
+      deleteByIdPreparedStatement2=null;
     }
   }
 }
