+++ KriptonAndroidLibrary/src/test/java/sqlite/feature/many2many/CityDaoImpl.java
@@ -23,6 +23,8 @@ import java.util.List;
 public class CityDaoImpl extends AbstractDao implements CityDao {
   private SQLiteStatement insertPreparedStatement0;
 
+  private SQLiteStatement deleteByIdPreparedStatement1;
+
   public CityDaoImpl(BindPersonCirtyDataSource dataSet) {
     super(dataSet);
   }
@@ -105,7 +107,7 @@ public class CityDaoImpl extends AbstractDao implements CityDao {
    */
   @Override
   public long insert(City bean) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     if (bean.name!=null) {
       _contentValues.put("name", bean.name);
     } else {
@@ -234,23 +236,26 @@ public class CityDaoImpl extends AbstractDao implements CityDao {
    */
   @Override
   public int deleteById(long id) {
-    KriptonContentValues _contentValues=contentValues();
+    KriptonContentValues _contentValues=contentValuesForUpdate();
     _contentValues.addWhereArgs(String.valueOf(id));
 
-    StringBuilder _sqlBuilder=getSQLStringBuilder();
     // generation CODE_001 -- BEGIN
     // generation CODE_001 -- END
+    if (deleteByIdPreparedStatement1==null) {
+      StringBuilder _sqlBuilder=getSQLStringBuilder();
 
-    // manage WHERE arguments -- BEGIN
+      // manage WHERE arguments -- BEGIN
 
-    // manage WHERE statement
-    String _sqlWhereStatement=" id=?";
-    _sqlBuilder.append(_sqlWhereStatement);
+      // manage WHERE statement
+      String _sqlWhereStatement=" id=?";
+      _sqlBuilder.append(_sqlWhereStatement);
 
-    // manage WHERE arguments -- END
+      // manage WHERE arguments -- END
 
-    // generate sql
-    String _sql="DELETE FROM cities WHERE id=?";
+      // generate sql
+      String _sql="DELETE FROM cities WHERE id=?";
+      deleteByIdPreparedStatement1 = KriptonDatabaseWrapper.compile(dataSource, _sql);
+    }
 
     // display log
     Logger.info("DELETE FROM cities WHERE id=?");
@@ -261,7 +266,7 @@ public class CityDaoImpl extends AbstractDao implements CityDao {
       Logger.info("==> param%s: '%s'",(_whereParamCounter++), StringUtils.checkSize(_whereParamItem));
     }
     // log for where parameters -- END
-    int result = KriptonDatabaseWrapper.updateDelete(dataSource, _sql, _contentValues);
+    int result = KriptonDatabaseWrapper.updateDelete(dataSource, deleteByIdPreparedStatement1, _contentValues);
     return result;
   }
 
@@ -269,6 +274,10 @@ public class CityDaoImpl extends AbstractDao implements CityDao {
     if (insertPreparedStatement0!=null) {
       insertPreparedStatement0.close();
       insertPreparedStatement0=null;
+    }
+    if (deleteByIdPreparedStatement1!=null) {
+      deleteByIdPreparedStatement1.close();
+      deleteByIdPreparedStatement1=null;
     }
   }
 }
